+++ Sources/SuperwallKit/Paywall/Presentation/PaywallInfo.swift
@@ -108,6 +108,12 @@ public final class PaywallInfo: NSObject {
   /// The time it took to load the paywall products.
   public let productsLoadDuration: TimeInterval?
 
+  /// An iso date string indicating when the shimmer view began loading.
+  public let shimmerLoadStartTime: String?
+
+  /// An iso date string indicating when the shimmer view finished loading.
+  public let shimmerLoadCompleteTime: String?
+
   /// The paywall.js version installed on the paywall website.
   public let paywalljsVersion: String?
 
@@ -134,6 +140,9 @@ public final class PaywallInfo: NSObject {
   /// Information about the presentation of the paywall.
   public let presentation: PaywallPresentationInfo
 
+  /// Indicates whether scrolling of the webview is enabled.
+  public let isScrollEnabled: Bool
+
   init(
     databaseId: String,
     identifier: String,
@@ -154,6 +163,8 @@ public final class PaywallInfo: NSObject {
     productsLoadStartTime: Date?,
     productsLoadFailTime: Date?,
     productsLoadCompleteTime: Date?,
+    shimmerLoadStartTime: Date?,
+    shimmerLoadCompleteTime: Date?,
     experiment: Experiment?,
     paywalljsVersion: String?,
     isFreeTrialAvailable: Bool,
@@ -163,7 +174,8 @@ public final class PaywallInfo: NSObject {
     localNotifications: [LocalNotification],
     computedPropertyRequests: [ComputedPropertyRequest],
     surveys: [Survey],
-    presentation: PaywallPresentationInfo
+    presentation: PaywallPresentationInfo,
+    isScrollEnabled: Bool
   ) {
     self.databaseId = databaseId
     self.identifier = identifier
@@ -226,7 +238,12 @@ public final class PaywallInfo: NSObject {
     } else {
       self.productsLoadDuration = nil
     }
+
+    self.shimmerLoadStartTime = shimmerLoadStartTime?.isoString ?? ""
+    self.shimmerLoadCompleteTime = shimmerLoadCompleteTime?.isoString ?? ""
+
     self.closeReason = closeReason
+    self.isScrollEnabled = isScrollEnabled
   }
 
   func eventParams(
@@ -252,13 +269,16 @@ public final class PaywallInfo: NSObject {
       "paywall_products_load_complete_time": productsLoadCompleteTime as Any,
       "paywall_products_load_fail_time": productsLoadFailTime as Any,
       "paywall_products_load_duration": productsLoadDuration as Any,
+      "shimmerView_load_complete_time": shimmerLoadCompleteTime as Any,
+      "shimmerView_load_start_time": shimmerLoadStartTime as Any,
       // TODO: Remove in v4:
       "trigger_session_id": "" as Any,
       "experiment_id": experiment?.id as Any,
       "variant_id": experiment?.variant.id as Any,
       "cache_key": cacheKey,
       "build_id": buildId,
-      "close_reason": closeReason.description
+      "close_reason": closeReason.description,
+      "is_scroll_enabled": isScrollEnabled as Any
     ]
 
     var loadingVars: [String: Any] = [:]
@@ -352,6 +372,8 @@ extension PaywallInfo: Stubbable {
       productsLoadStartTime: nil,
       productsLoadFailTime: nil,
       productsLoadCompleteTime: nil,
+      shimmerLoadStartTime: nil,
+      shimmerLoadCompleteTime: nil,
       experiment: nil,
       paywalljsVersion: nil,
       isFreeTrialAvailable: false,
@@ -365,7 +387,8 @@ extension PaywallInfo: Stubbable {
         style: .none,
         condition: .checkUserSubscription,
         delay: 0
-      )
+      ),
+      isScrollEnabled: true
     )
   }
 
@@ -391,6 +414,8 @@ extension PaywallInfo: Stubbable {
       productsLoadStartTime: nil,
       productsLoadFailTime: nil,
       productsLoadCompleteTime: nil,
+      shimmerLoadStartTime: nil,
+      shimmerLoadCompleteTime: nil,
       experiment: .init(
         id: "0",
         groupId: "0",
@@ -412,7 +437,8 @@ extension PaywallInfo: Stubbable {
         style: .none,
         condition: .checkUserSubscription,
         delay: 0
-      )
+      ),
+      isScrollEnabled: true
     )
   }
 }