+++ Sources/SuperwallKit/StoreKit/Transactions/TransactionManager.swift
@@ -70,6 +70,14 @@ final class TransactionManager {
 
     let result = await purchase(product, purchaseSource: purchaseSource)
 
+    // Return early if using a purchase controller and purchasing externally.
+    // This avoids duplicate calls by the purchase function of the purchase
+    // controller.
+    if case .external = purchaseSource,
+      factory.makeHasExternalPurchaseController() {
+      return result
+    }
+
     switch result {
     case .purchased:
       await didPurchase(
@@ -275,7 +283,9 @@ final class TransactionManager {
         state: .restore(restoreType),
         paywallInfo: paywallInfo,
         product: product,
-        model: transaction
+        model: transaction,
+        source: .internal,
+        storeKitVersion: .storeKit1
       )
       await Superwall.shared.track(trackedEvent)
       await paywallViewController.webView.messageHandler.handle(.transactionRestore)
@@ -289,7 +299,9 @@ final class TransactionManager {
         state: .restore(restoreType),
         paywallInfo: .empty(),
         product: product,
-        model: transaction
+        model: transaction,
+        source: .external,
+        storeKitVersion: .storeKit1
       )
       await Superwall.shared.track(trackedEvent)
     }
@@ -303,7 +315,8 @@ final class TransactionManager {
       return .failed(PurchaseError.productUnavailable)
     }
     await factory.makePurchasingCoordinator().beginPurchase(
-      of: product.productIdentifier
+      of: product.productIdentifier,
+      source: purchaseSource
     )
     switch purchaseSource {
     case .internal:
@@ -343,7 +356,9 @@ final class TransactionManager {
           state: .fail(.failure(error.safeLocalizedDescription, product)),
           paywallInfo: paywallInfo,
           product: product,
-          model: nil
+          model: nil,
+          source: .internal,
+          storeKitVersion: .storeKit1
         )
         await Superwall.shared.track(trackedEvent)
         await paywallViewController.webView.messageHandler.handle(.transactionFail)
@@ -364,7 +379,9 @@ final class TransactionManager {
           state: .fail(.failure(error.safeLocalizedDescription, product)),
           paywallInfo: .empty(),
           product: product,
-          model: nil
+          model: nil,
+          source: .external,
+          storeKitVersion: .storeKit1
         )
         await Superwall.shared.track(trackedEvent)
       }
@@ -392,7 +409,9 @@ final class TransactionManager {
         state: .start(product),
         paywallInfo: paywallInfo,
         product: product,
-        model: nil
+        model: nil,
+        source: .internal,
+        storeKitVersion: .storeKit1
       )
       await Superwall.shared.track(trackedEvent)
       await paywallViewController.webView.messageHandler.handle(.transactionStart)
@@ -401,6 +420,11 @@ final class TransactionManager {
         paywallViewController.loadingState = .loadingPurchase
       }
     case .external:
+      // If an external purchase controller is being used, skip because this will
+      // get called by the purchase function of the purchase controller.
+      if factory.makeHasExternalPurchaseController() {
+        return
+      }
       Logger.debug(
         logLevel: .debug,
         scope: .paywallTransactions,
@@ -411,7 +435,9 @@ final class TransactionManager {
         state: .start(product),
         paywallInfo: .empty(),
         product: product,
-        model: nil
+        model: nil,
+        source: .external,
+        storeKitVersion: .storeKit1
       )
       await Superwall.shared.track(trackedEvent)
     }
@@ -506,7 +532,9 @@ final class TransactionManager {
         state: .abandon(product),
         paywallInfo: paywallInfo,
         product: product,
-        model: nil
+        model: nil,
+        source: .internal,
+        storeKitVersion: .storeKit1
       )
       await Superwall.shared.track(trackedEvent)
       await paywallViewController.webView.messageHandler.handle(.transactionAbandon)
@@ -527,7 +555,9 @@ final class TransactionManager {
         state: .abandon(product),
         paywallInfo: .empty(),
         product: product,
-        model: nil
+        model: nil,
+        source: .external,
+        storeKitVersion: .storeKit1
       )
       await Superwall.shared.track(trackedEvent)
     }
@@ -550,7 +580,9 @@ final class TransactionManager {
         state: .fail(.pending("Needs parental approval")),
         paywallInfo: paywallInfo,
         product: nil,
-        model: nil
+        model: nil,
+        source: .internal,
+        storeKitVersion: .storeKit1
       )
       await Superwall.shared.track(trackedEvent)
       await paywallViewController.webView.messageHandler.handle(.transactionFail)
@@ -566,7 +598,9 @@ final class TransactionManager {
         state: .fail(.pending("Needs parental approval")),
         paywallInfo: .empty(),
         product: nil,
-        model: nil
+        model: nil,
+        source: .external,
+        storeKitVersion: .storeKit1
       )
       await Superwall.shared.track(trackedEvent)
     }
@@ -627,7 +661,9 @@ final class TransactionManager {
         state: .complete(product, transaction),
         paywallInfo: paywallInfo,
         product: product,
-        model: transaction
+        model: transaction,
+        source: .internal,
+        storeKitVersion: .storeKit1
       )
       await Superwall.shared.track(trackedEvent)
       await paywallViewController.webView.messageHandler.handle(.transactionComplete)
@@ -667,7 +703,9 @@ final class TransactionManager {
         state: .complete(product, transaction),
         paywallInfo: .empty(),
         product: product,
-        model: transaction
+        model: transaction,
+        source: .external,
+        storeKitVersion: .storeKit1
       )
       await Superwall.shared.track(trackedEvent)